+++ config/services/tap.mjs
@@ -27,11 +27,12 @@ export const resolveServiceConfiguration = ({ utils }) => {
       volumes: PROD
         ? [
             `${utils.getPreset('MORIO_CONFIG_ROOT')}/tap:/morio/tap/config`,
-            `${utils.getPreset('MORIO_CONFIG_ROOT')}/tap/processors:/morio/tap/processors`,
+            `${utils.getPreset('MORIO_CONFIG_ROOT')}/shared/processors:/morio/tap/processors`,
           ]
         : [
             `${utils.getPreset('MORIO_GIT_ROOT')}:/morio`,
             `${utils.getPreset('MORIO_GIT_ROOT')}/data/config/tap:/morio/tap/config`,
+            `${utils.getPreset('MORIO_GIT_ROOT')}/data/config/shared/processors:/morio/tap/processors`,
           ],
     },
     /*